+++ example/art_pen/gen/art/pen/entries.dart
@@ -17,7 +17,7 @@ class PenEntries extends ModelEntries {
   Entities newEntities(String conceptCode) {
     var concept = model.concepts.singleWhereCode(conceptCode);
     if (concept == null) {
-      throw new ConceptError("${conceptCode} concept does not exist.") ;
+      throw new ConceptException("${conceptCode} concept does not exist.") ;
     }
     if (concept.code == "Segment") {
       return new Segments(concept);
@@ -31,7 +31,7 @@ class PenEntries extends ModelEntries {
   ConceptEntity newEntity(String conceptCode) {
     var concept = model.concepts.singleWhereCode(conceptCode);
     if (concept == null) {
-      throw new ConceptError("${conceptCode} concept does not exist.") ;
+      throw new ConceptException("${conceptCode} concept does not exist.") ;
     }
     if (concept.code == "Segment") {
       return new Segment(concept);