+++ example/category_links/gen/category/links/entries.dart
@@ -23,7 +23,7 @@ class LinksEntries extends ModelEntries {
   Entities newEntities(String conceptCode) { 
     var concept = model.concepts.singleWhereCode(conceptCode); 
     if (concept == null) { 
-      throw new ConceptError("${conceptCode} concept does not exist.") ; 
+      throw new ConceptException("${conceptCode} concept does not exist.") ; 
     } 
     if (concept.code == "Member") { 
       return new Members(concept); 
@@ -49,7 +49,7 @@ class LinksEntries extends ModelEntries {
   ConceptEntity newEntity(String conceptCode) { 
     var concept = model.concepts.singleWhereCode(conceptCode); 
     if (concept == null) { 
-      throw new ConceptError("${conceptCode} concept does not exist.") ; 
+      throw new ConceptException("${conceptCode} concept does not exist.") ; 
     } 
     if (concept.code == "Member") { 
       return new Member(concept);