+++ example/dartling_types/gen/dartling/types/entries.dart
@@ -17,7 +17,7 @@ class TypesEntries extends ModelEntries {
   Entities newEntities(String conceptCode) { 
     var concept = model.concepts.singleWhereCode(conceptCode); 
     if (concept == null) { 
-      throw new ConceptError("${conceptCode} concept does not exist.") ; 
+      throw new ConceptException("${conceptCode} concept does not exist.") ; 
     } 
     if (concept.code == "Type") { 
       return new Types(concept); 
@@ -28,7 +28,7 @@ class TypesEntries extends ModelEntries {
   ConceptEntity newEntity(String conceptCode) { 
     var concept = model.concepts.singleWhereCode(conceptCode); 
     if (concept == null) { 
-      throw new ConceptError("${conceptCode} concept does not exist.") ; 
+      throw new ConceptException("${conceptCode} concept does not exist.") ; 
     } 
     if (concept.code == "Type") { 
       return new Type(concept);