+++ example/game_parking/gen/game/parking/entries.dart
@@ -21,7 +21,7 @@ class ParkingEntries extends ModelEntries {
   Entities newEntities(String conceptCode) { 
     var concept = model.concepts.singleWhereCode(conceptCode); 
     if (concept == null) { 
-      throw new ConceptError("${conceptCode} concept does not exist.") ; 
+      throw new ConceptException("${conceptCode} concept does not exist.") ; 
     } 
     if (concept.code == "Area") { 
       return new Areas(concept); 
@@ -41,7 +41,7 @@ class ParkingEntries extends ModelEntries {
   ConceptEntity newEntity(String conceptCode) { 
     var concept = model.concepts.singleWhereCode(conceptCode); 
     if (concept == null) { 
-      throw new ConceptError("${conceptCode} concept does not exist.") ; 
+      throw new ConceptException("${conceptCode} concept does not exist.") ; 
     } 
     if (concept.code == "Area") { 
       return new Area(concept);