+++ example/todo_mvc/todo/mvc/tasks.dart
@@ -25,7 +25,7 @@ class Tasks extends TasksGen {
  
   // added after code gen - begin 
 
-  Tasks get completed => selectWhere((task) => task.completed);
+  Tasks get completed => selectWhere((task) => task.isCompleted);
   Tasks get left => selectWhere((task) => task.left);
 
   bool preAdd(Task task) {