+++ lib/dartling.dart
@@ -55,7 +55,7 @@ part 'gen/search.dart';
 part 'domain/model/event/actions.dart';
 part 'domain/model/event/past.dart';
 part 'domain/model/event/reactions.dart';
-part 'domain/model/error/errors.dart';
+part 'domain/model/error/exceptions.dart';
 part 'domain/model/error/validations.dart';
 part 'domain/model/transfer/json.dart';
 part 'domain/model/entities.dart';