+++ lib/domain/model/id.dart
@@ -191,7 +191,7 @@ class Id implements IdApi {
        }
        return compare;
      }
-     throw new IdError('${_concept.code}.id does not have parents.');
+     throw new IdException('${_concept.code}.id does not have parents.');
    }
 
    /**
@@ -215,7 +215,7 @@ class Id implements IdApi {
        } // for
        return compare;       
      }  
-     throw new IdError('${_concept.code}.id does not have attributes.');
+     throw new IdException('${_concept.code}.id does not have attributes.');
    }
 
    /**
@@ -235,7 +235,7 @@ class Id implements IdApi {
       }
       return compare;
     }
-    throw new IdError('${_concept.code}.id is not defined.');
+    throw new IdException('${_concept.code}.id is not defined.');
   }
 
   String _dropEnd(String text, String end) {