+++ lib/meta/concepts.dart
@@ -339,7 +339,7 @@ class Concept extends ConceptEntity<Concept> {
           return parent.destinationConcept.entryConcept;
         }
       }
-      throw new ParentError('No internal parent for the ${code} concept');
+      throw new ParentException('No internal parent for the ${code} concept');
     }
   }
 
@@ -354,7 +354,7 @@ class Concept extends ConceptEntity<Concept> {
             '${code}';
         }
       }
-      throw new ParentError('No internal parent for the ${code} concept');
+      throw new ParentException('No internal parent for the ${code} concept');
     }
   }